Move resolveProxy From app to session
